backend/requirements.txt
fastapi
uvicorn
librosa
pydub
pedalboard
numpy
python-multipart
soundfile
requests
oss2
redis
rq
demucs
torch
torchaudio
python-dotenv
